# mobile/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/classrooms/{classroomId}/activity/{activityId}": {
      "get": {
        "operationId": "ActivityController_findOne",
        "parameters": [
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "JWT-auth": [] }],
        "summary": "Get Activity by id",
        "tags": ["Activities"]
      },
      "patch": {
        "operationId": "ActivityController_update",
        "parameters": [
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateActivityDTO" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "security": [{ "JWT-auth": [] }],
        "summary": "Update Activity",
        "tags": ["Activities"]
      },
      "delete": {
        "operationId": "ActivityController_remove",
        "parameters": [
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "JWT-auth": [] }],
        "summary": "Delete Activity",
        "tags": ["Activities"]
      }
    },
...
    "/classroom/activity-logs/{activityId}": {
      "post": {
        "operationId": "ActivityLogsController_create",
        "parameters": [
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateActivityLogDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "JWT-auth": [] }],
        "summary": "Create Activity Log",
        "tags": ["ActivityLogs"]
      },
      "get": {
        "operationId": "ActivityLogsController_findOne",
        "parameters": [
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "JWT-auth": [] }],
        "summary": "Get an Activity's Activity Logs",
        "tags": ["ActivityLogs"]
      }
    },
    "/classroom/activity-logs/{activityId}/classroom/{classroomId}/activity-logs": {
      "get": {
        "operationId": "ActivityLogsController_findAll",
        "parameters": [
          {
            "name": "classroomId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "JWT-auth": [] }],
        "summary": "Get all Classroom Acitivies' Activity Log",
        "tags": ["ActivityLogs"]
      }
    },
...
  },
  "info": {
    "title": "LexiLearner API",
    "description": "",
    "version": "1.0",
    "contact": {},
    "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" }
  },
...
  "servers": [
    { "url": "http://localhost:3000", "description": "Development server" }
  ],
  "components": {
    "schemas": {
...
      "UpdateActivityDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Activity Name",
            "example": "Cat in the Hat Activity"
          },
          "description": {
            "type": "string",
            "description": "Description on what the activity is about",
            "example": "Read the Cat in the Hat book and understand the adventures of our beloved main characters!"
          },
          "reading_material_id": { "type": "string" }
        }
      },
      "CreateActivityLogDto": {
        "type": "object",
        "properties": {
          "minigame_log_id": {
            "type": "string",
            "description": "Minigame log Id"
          }
        },
        "required": ["minigame_log_id"]
      }
    }
  }
}